refactor-864370: Add addItems() API
Having [add|remove][Command|Type|Converter|etc] is nice, but we often
want to register a bunch of things so we add a module that pulls
together all the addFoo()/removeFoo() commands, and adds a addItems()
and removeItems().

A command, converter, etc declares what type of thing it is using the
'item' property which should be one of 'command', 'setting', 'field',
'type' or 'converter'.

Signed-off-by: Joe Walker <jwalker@mozilla.com>