refactor-864370: s/skipArgUpdate/internal
Requisition.setAssignment had a skipArgUpdate option which obviously
skipped the 'arg update' process. This processing is needed when
setAssignment is being called from outside of a Requisition, but not
when it is an internal part of a call to update(). Since we are about
to add other processing with the same internal/extenal switch it seems
sensible to rename the option.

Signed-off-by: Joe Walker <jwalker@mozilla.com>