refactor-864370: Introduce updateIds to Requisition
Calls to update() and setAssignment() happen asynchronously, which means
we could call update twice before the first has completed. This change
adds an updateId which tracks a count of the changes so we can bail out
if we notice that the current change id is greater than the one we're
working on.

This (partially, see the next changeset) fixes an old bug, which wasn't
really noticeable until file types brought genuinely slow parsing.

Signed-off-by: Joe Walker <jwalker@mozilla.com>