refactor-864370: Upgrade Event.holdFire
GCLI's EventEmitter has a holdFire function which is useful when you're
doing a bunch of actions all of which would normally fire events, but
you want to coalesce them into a single event firing.

The prior implementation used a 'holdFire' boolean to achieve this, but
that means that calls to event.holdFire() and event.resumeFire() can't
be nested, so we replace 'holdFire' with an integer 'fireHoldCount'.

Signed-off-by: Joe Walker <jwalker@mozilla.com>