output-901602: Have Output.promise resolve with itself
Previously there was a nasty thing where requisition.exec() returned an
output object, directly. output.promise resolved when the command was
complete, but with undefined, and the data stored on the output object.
This was for historical reasons, but now output.promise resolves with
itself, and all uses of requisition.exec() use the output object using
the 'then' function rather than directly in the return value.

Signed-off-by: Joe Walker <jwalker@mozilla.com>