output-901602: Don't reject output on error - resolve
Previously when a command failed it rejected the promise, since the
bot reject and resolve both used an Output object, and the Output object
contained an error field we don't need reject, and it's actually
annoying because we generally want to do that same thing (i.e. display
the data) both ways.

Signed-off-by: Joe Walker <jwalker@mozilla.com>