asynctests-904549: Keep in sync with m-c helpers.js
If options.window is null then it's likely that helpers.audit has been
called and the returned promise ignored, so the test has finished before
all the results are in. This adds a warning and attempts to fail the
test in these cases.

This change had already been reviewed in
https://bugzilla.mozilla.org/show_bug.cgi?id=904549

Signed-off-by: Joe Walker <jwalker@mozilla.com>