nodeinout-796921: Update Highlighter for Firefox
The Firefox developer tools Highlighter rewrite is incomplete, so node
highlighting will not work yet. GCLI has a Highligher, and the web
interface changed.
This change updates the Firefox version of the GCLI Highligher proxy to
match the web version.
When the developer tools Highlighter rewrite is done, we'll need to
implement _highlightNode and _unhighlightNode to call into that Firefox
code.

Signed-off-by: Joe Walker <jwalker@mozilla.com>