jsterm-911456: Add stringView converter
Previously when we wanted to format something for string output (e.g.
when in NodeJS) we needed to inject a custom template which rendered to
string better. This made the output style something that was selected by
the configuration. This change adds a new stringView converter that
allows separate converters for string and dom output.

We use this to move the custom templates for help from server to gcli
which makes the help command usable from JSTerm.

Signed-off-by: Joe Walker <jwalker@mozilla.com>