easyexec-767340: Export Requisition from gcli.jsm
Requisition was not exported from gcli.jsm because it's not considered
a 'stable' API, but we needed to use it a lot and it's hard to get at.

While it's not stable in that it is likely to change it's not too
unstable for use within Firefox developer tools, so it is OK to have
it exported from the JSM. Clearly difficulty to access isn't stopping us
accessing it when needed.

Signed-off-by: Joe Walker <jwalker@mozilla.com>