refactor-864370: Use 'rank' to describe a chosen prediction
getPredictionAt() was confusing because at implies a location, and this
is just about which of the predictions has been chosen. Worse there is
a location (that of the cursor) to confuse things. Worse still, we had
a number of words to describe this: 'at', 'index', 'predictionChoice',
and 'choice'. Now we use 'rank' for everything.

Signed-off-by: Joe Walker <jwalker@mozilla.com>