jsterm-911456: getStateData should return a promise
Previously it returned a object that contained both read data and
promises. All the promises were waiting for the same thing so we could
easily return a single promise instead.

We're doing this because we need to remote this function so the data
that it returns needs to be JSON or a promise of JSON, and there isn't
a great upside to the more complex way of doing things - only really
that we could display input markup directly before we knew about
predictions.

Signed-off-by: Joe Walker <jwalker@mozilla.com>