refactor-864370: Enable double click for help synopsis
Previously the synopsis for help output was clickable (copy to input)
but not double-clickable (execute) like most other commands because
most commands can't be executed without parameters. But this is annoying
for those that can, so we make updateExec only execute when the status
is valid to catch the cases when parameters are needed, and enable the
double click hook.

Signed-off-by: Joe Walker <jwalker@mozilla.com>