refactor-864370: More descriptive error messages
When executing a command using the API (rather than the GUI) it is
possible to execute a command with incomplete parameters, which
previously didn't have any error message associated with them. This
adds an error message for incomplete parameters.

Many tests checked to see if the message was empty - mostly these checks
are not needed (we generally don't care about messages that we are not
going to display in the general case) so the majority of these checks
have been removed, with just a few left in testCli.js

Signed-off-by: Joe Walker <jwalker@mozilla.com>