jsterm-911456: Steps to remove JSDom
Test only changes.

I had a go at removing the need for JSDom. The first step is to ensure
that the tests will work from the command line without JSDom, so I've
added a NoDom option. This patch does the following:

* Adds checks for options.terminal to helpers and doesn't run the checks
  that require a terminal. Future updates should reduce the checks that
  are skipped, however many of the checks are directly on the UI and
  so can be safely skipped.
* Adds skip markers to code that won't run without a DOM.
* Updates the test command to have a "--nodom" option to skip the
  creation of a JSDom environment.

The next step is to remove the actual supporting JSDom code, however
before then we need to have a template mechanism that works on strings.

Signed-off-by: Joe Walker <jwalker@mozilla.com>