refactor-864370: Remove remaining JSHint warnings
Complaining about properties not being in camel case doesn't help
because we don't have a problem with that, and it gets in the way when
we do it for good reason.

3 places where we had things which were both variables and constructor
function followed the variable naming convention (camelCase, starts
lower) rather than the constructor naming convention (CamelCase,
starts upper) but JSHint doesn't like this and there's no strong reason
to prefer one over the other, so I changed the variable names to start
upper.

Renamed history to dirHistory to avoid conflict with globals

Signed-off-by: Joe Walker <jwalker@mozilla.com>