browserify-919553: Remove async requires from test/index.js
I'm not 100% clear why we needed async requires in test/index.js,
probably something to do with top level scopes and the old version of
RequireJS that we were using.

Since the upgrade this causes error messages, so we do the simpler
thing of using sync require like everything else.

Signed-off-by: Joe Walker <jwalker@mozilla.com>