jsterm-911456: Remove hint positioning
It seems likely that JS won't be able to use this assignment based way
of looking at hints (and moving horizontally with the cursor could be
really annoying).

Also we are looking about the hint area being significantly constricted
vertically, so we may wish to use all the horizontal space we can.

Signed-off-by: Joe Walker <jwalker@mozilla.com>