jsterm-911456: Remove close for OutputView
OutputView had a number of theoretical features like the ability to
hide and how output and remove it entirely. None of these was used in
any great way (close being a near exception, used to remove some test
command output, to be replaced by a 'clear' command)

For simplicities sake we're removing these features.

Signed-off-by: Joe Walker <jwalker@mozilla.com>