jsterm-911456: Create a 'clear' command
In the process of adding the clear command the following changes
happened:
Moved util.linksToNewTab to CommandControl because this was the only
place that it was used, so it feels a bit less like a utility.
Added clear function to terminal, and had terminal add itself to the
root element to allow things to find it and use the clear function.

Signed-off-by: Joe Walker <jwalker@mozilla.com>