jsterm-911456: Further tweaks to the look and feel
Remove comments from the template. Whilst helpful they clog up the page
due to the number of times they are included.
Force the prompt and output spans to be contiguous so we don't have to
factor a (potentially variable width) space into our alignment issues.
Remove hover CSS as it's no longer needed.
Change shortcut prompt to a ':" for the command system.

Signed-off-by: Joe Walker <jwalker@mozilla.com>