jsterm-911456: Exclude CSS for Firefox use
Trying to keep the same set of CSS rules between the browser and Firefox
is a good goal, but it just doesn't work. Sometimes we want it to look
different, sometimes Firefox doesn't define elements the same way (e.g.
pre) and sometimes there are different restrictions (e.g. font use). And
there is a different way of including files, so we're just going to
ignore CSS imports done the GCLI way, use an empty string and put the
CSS in conventional CSS places in Firefox.

Signed-off-by: Joe Walker <jwalker@mozilla.com>