refactor-864370: Pass a boolean into FocusManager.setError
The code in focus manager works so long as the error message isn't empty
but it is confusing. Also updated the comment to ignore an issue of
modularity (which doesn't seem so important with recent refactorings)
and point out a difference in logic which we don't think causes any
actual bugs right now, but which we should fix next time we think about
this code.

Signed-off-by: Joe Walker <jwalker@mozilla.com>