jsterm-911456: Make FFDisplay start synchronously
We made resource loading asynchronous (which generally makes sense)
however the display system for Firefox doesn't really need to be async
because the strings it wants to load are simple and arguably better
being inlined, and because the startup of the developer toolbar would
have to include another async step.

So this makes the startup of FFDisplay synchronous again, and in doing
so fixes the race in running tests.

Signed-off-by: Joe Walker <jwalker@mozilla.com>