syncincdec-830184: Make _assign properly asynchronous
Previously there was an ‘outstanding’ array in which we collected
promises of setAssignments that were not complete, and we finished with
‘promise.all(outstanding);’. However that allows the setAssignment calls
to complete in any order, which breaks delegated types which are allowed
to know about other assignments (which are now in an ambiguous state).
So we switch to using util.promiseEach which processes each pass of the
array in order before moving on to the next.

Signed-off-by: Joe Walker <jwalker@mozilla.com>