syncincdec-830184: Comment out tests depending on bug 935990
There are a couple of places where asynchronous output is used in a
template, however the domtemplate engine doesn’t make it easy for us to
know when these updates are completed. Since these issues don’t
represent real world bugs, we’re commenting the tests out until a fix
is ready.

Signed-off-by: Joe Walker <jwalker@mozilla.com>