xpcshellcrash-936451: Workaround XHR issues
Without this patch ‘mach package’ fails with a dump in xpcshell.

xpcshell is computing a startup cache by loading each of the JS files in
turn and running a single pass through the event loop on each to see
what things they load. Recent changes in GCLI split out a large file
(gcli.jsm) into a number of smaller files, some of which use other
resources fetched using an XHR request to a resource://... URL.

It looks like there is some sort of security principle which is normally
setup for a working Firefox, which isn't there in the xpcshell
environment, so when XHR tries to find out if it's allowed to read from
the resource://... URL, it dumps on an null pointer.

So there's an easy workaround, I simply setTimeout before doing the XHR
so the call doesn't happen on the first pass through the event loop.
This somewhat subverts the whole point of this process, however since
these files should not be parsed on startup, I'm not sure what the
performance impact will be.

Signed-off-by: Joe Walker <jwalker@mozilla.com>