refactor-864370: Minor - Update test commands
Steps toward making the 2 test commands be the same. They are different
because the nodejs version uses JS DOM, but we’re working towards
getting rid of that version because it causes too many hacks in our
code.

These changes mostly make the 2 code paths similar and introduce --nodom
testing to the web version of the code.

Signed-off-by: Joe Walker <jwalker@mozilla.com>