xpcshellcrash-936451: Workaround XHR issues (attempt 2)
This undoes our previous 2 attempts at fixing the xpcshell crash and
replaces them with another. The l10n fix (to intro.js and focus.js)
prevented some error messages, but didn’t touch the real issue and the
workaround is nasty enough that I’m removing it.

The setTimeout-before-XHR solution just moved the problem later on
(because xpcshell noticed a hanging task) so the revised solution is to
only load the resource when we are actually going to use them.

Signed-off-by: Joe Walker <jwalker@mozilla.com>