jsterm-911456: Automatically add/remote requisition mappings
Previously things that created a Requisition were responsible for adding
and removing the mapping, but this change moves this responsibility
from the creator to the Requisition itself.

Requisition is the model behind a command line, but it’s API isn’t
designed for external use, so we have conversionContext and
executionContext which are designed to be passed to commands and types.
But there are times when bits of the guts of GCLI need to get at a
Requisition, when all they have is a context. So we had a mapping to
allow them to get at the Requisition.

Signed-off-by: Joe Walker <jwalker@mozilla.com>