refactor-864370: Pass a context to getPredictions() (part 2)
Now that we’re passing a context into getPredictions(), we pass this on
use this context in SelectionType.getBlank rather than the context
that was sometimes passed to getBlank() and sometimes not. We also
simplify the Converter ctor so we don’t call the prediction function
just to check that’s it works properly (as we now need a context to do
this anyway).

Signed-off-by: Joe Walker <jwalker@mozilla.com>