jsterm-911456: Simplify startup
A number of changes aimed at simplifying the code paths at startup.

Merge localtest.html and index.html. This removes some duplication and
the greater complexity in index.html is useful documentation anyway,
and it makes the test case startup much simpler.

Move to a model of registering items in groups. It is hoped that we will
be able to have some standard lists of items to remove the duplication
of the lists for different startup cases.

Make commandSpecs and typeSpecs be arrays, so they can more directly
be used in a call to addItems.

Add a ‘direct’ connector to help debugging.

Signed-off-by: Joe Walker <jwalker@mozilla.com>