orange-938575: Avoid intermittent orange in Firefox
Bug 938575 [1] was raised after a merge [2] which resulted in a
changeset [3] which seems to have nothing to do with GCLI, but which
caused an intermittent orange on GCLI tests when run under Fedora 12.
This fix is a temporary measure which removes the parts of the test that
caused this problem. This is something of a temporary band-aid solution
because we don’t understand the root cause yet, however we do want to
stop the oranges.

A fuller fix for this problem is likely to be low priority because this
scenario is tested in many unit tests and in many environments without
failure.

[1]: https://bugzilla.mozilla.org/show_bug.cgi?id=938575
[2]: https://hg.mozilla.org/mozilla-central/pushloghtml?startID=25652&endID=25653
[3]: https://hg.mozilla.org/mozilla-central/raw-rev/319c3d822ae4

Signed-off-by: Joe Walker <jwalker@mozilla.com>