jsterm-911456: Move mockCommand setup to top level
Previously each individual test called mockCommands.setup/shutdown as
part of the setup/shutdown for that test, but the command setup for
remote tests is harder (the commands need to be registered on the server
and then imported to the client) so we needed the setup to be handled
at the top level.

The logic behind the old system was that the Firefox GCLI test runner
wouldn’t need to know about the mock commands (that was the
responsibility of the test after all) so we also need to update the code
that converts GCLI tests to mochi tests to add the mockCommand.setup/
shutdown code.

Signed-off-by: Joe Walker <jwalker@mozilla.com>