refactor-864370: Move types/canon/history to more logical homes
3 files (canon/types/argument) were there to manage the 2 basic
extension points (commands and types). We store other similar management
code in with the things being managed, so it makes sense to do the same
here. history.js is really UI code, so we move that into the ui module
too.
So this changeset, although it looks large is just a bunch of tweaked
require statements.

Signed-off-by: Joe Walker <jwalker@mozilla.com>