jsterm-911456: Unify the lists of registered items
We had several ways to register items, and were working towards getting
them all in line. This finishes the job. Now all the lists of items
are the same bad commenting out where an item shouldn't be used in a
particular context.
This includes removing mozindex.js which isn't actually used.

Signed-off-by: Joe Walker <jwalker@mozilla.com>