srcdir-979030: Fork the mozilla branch
Previously we managed alternate implementations using loader tricks
but this was getting increasingly messy, so now we're going to do it
using a git branch. This means quite a bit of moving files around, but
no substantive changes.

As far as the GCLI/master branch goes, it means that we can delete all
the mozilla specific files. This is the first commit to the mozmaster
branch which moves those mozilla specific files over their main branch
versions, and removes all the files that were not relevant to mozilla.

Signed-off-by: Joe Walker <jwalker@mozilla.com>