spawn-1007006: Rename host.exec to host.spawn
host.exec is a poor name because it spawns a new process, and we'd like
to have a host.exec function that does a JS exec, so first we rename
host.exec to host.spawn.

The slightly confusing thing is that the Firefox version of the new
host.exec is going to use Task.spawn. However this change just renames
a function. The Task.spawn thing will happen in a subsequent change.

Signed-off-by: Joe Walker <jwalker@mozilla.com>