toolkitpromise-992309: Fix text mode help
This is not strictly part of fixing bug 992309, but the changes to m-c
are trivial, so I'm including it there.

'gcli help' failed on the command line because there wasn't a good way
to generate a DOM to format the string output from the help command.

This moves the toDom function from util to host, so we can have a tweak
to use jsdom when there is no document available.

Signed-off-by: Joe Walker <jwalker@mozilla.com>