union-1016578: Use correct type store
centralTypes is the default to save implementors creating their own, but
we shouldn't assume the default. The standard way to access the current
types object is to use this.types, which is injected at creation time.
However initially this was restricted to selection/delegate types, however
it didn't make sense to retain this 'special status' so we've opened that
tweak to all types.

Signed-off-by: Joe Walker <jwalker@mozilla.com>