runat-1128988: Use system rather than 'gcli'
In many cases there is a gcli global that is a system object. We want
to make it clearer what we're working with and to get rid of the global

The changes to cli.js rid us of a useless function and make it clearer
how system is used inside a Requisition.

Signed-off-by: Joe Walker <jwalker@mozilla.com>