runat-1128988: Use a blank type when we can't resolve the delegate type
Delegated types are used when we're not sure a load time what the type
should be. An example is the 'pref set' command where the new pref
value is determined by the pref being set.

The addition of remoting made this asynchronous, but we've not had time
to make all the required adjustments yet. This just assumes a 'blank'
type if we can't ask.

Signed-off-by: Joe Walker <jwalker@mozilla.com>