runat-1128988: Add 'location' support to filter commands
Adding a 'location' to a system makes it ignore commands that don't have
a matching runAt property. This is principly for client/server setups
where we import commands from the server to the client, so a system with
`{ location: 'client' }` will silently ignore commands with
`{ runAt: 'server' }`. Any system without a location will accept
commands with any runAt property (including none).

Signed-off-by: Joe Walker <jwalker@mozilla.com>