runat-1128988: Improve support for custom properties for commands
GCLI has allowed (by turning a blind eye) to commands and other items
that have additional properties. This is used by Firefox to allow
commands to be placed in the toolbox and toolbar buttons with custom
icons.

With remote GCLI it's important that these custom properties are
transferred along with the standard command metadata. This change allows
the 'specs' function to take an array of custom properties which should
be transferred.

Signed-off-by: Joe Walker <jwalker@mozilla.com>