runat-1128988: Command specs that want to use fronts need to be objects
The server sends a block of JSON to the client describing the commands
and on the client the parameter types are turned into proper Type
objects and the 'front' object injected, but this can only happen if
the spec is an object rather than just a string.

The remote type should be able to remote itself trivially.

Signed-off-by: Joe Walker <jwalker@mozilla.com>