runat-1128988: Strengthen the tests we can make using mockCommands
Most of the test commands had an exec function generated by 'createExec'
which previously outputted (synchronously) a block of text that
described the arguments (using toString)

This had the problem that we couldn't test the values individually, so
now createExec returns a testCommandOutput structure containing the
correct argument stringified version (when promises).

A number of tests change to reflect this, and it needs an easy way to
lookup parameters by name in Command.

Signed-off-by: Joe Walker <jwalker@mozilla.com>