runat-1128988: Mock 'document' better
We had various hacky ways to mock and setup the document against which
the 'node' and 'resource' types acted, so we could run tests against
NodeJS, the web, the web remoted to NodeJS and Firefox. With this
change, all tests assume that the environment is setup correctly so
there is some sort of a DOM. We remove the hacky fake DOM and replace
it with a less hacky DOM from jsdom.

Signed-off-by: Joe Walker <jwalker@mozilla.com>