runat-1128988: Fix 'node' so it works properly remotely
This is mostly tweaks to tests. The node type doesn't call
createEmptyNodeList where it doesn't need to, and the NodeList type
now sets up the ability to convert back to a string properly so the
remote tests can at least prove that the value (which isn't available
remotely) could be stringified into something expected.

Signed-off-by: Joe Walker <jwalker@mozilla.com>