runat-1128988: Remove isNoDom
Several tests were skipped when there wasn't a DOM to test against.
This removes all those test skips. Given the recent fixes, most of the
skips could just be removed. Some of the skips are not down to
weaknesses in jsdom, but in NodeJS or the requisitionAutomator which is
used only in NodeJS, so some of this is s/isNoDom/isNode. Some of the
problems were confusion as to an id in the document that we were
testing against. #gcli-input was removed some time ago, so we're now
using #gcli-root which still exists. Finally I added a trivial
converter for the 'pref list' command so it would work on the command
line, and be tested properly.

Signed-off-by: Joe Walker <jwalker@mozilla.com>