runat-1128988: Replace document/globalObject with windowHolder
Previously each time we navigated to a new page we'd have to update the
document/global in GCLI. A windowHolder which uses a getter property to
dynamically return the correct document/global is much easier to use.

We're using windowHolder rather than documentHolder simply because you
could confuse a documentHolder and a window, but you can't confuse a
windowHolder and a document so easily.

The javascript type isn't used in Firefox. It wasn't clear that the
donteval test was working properly, and I didn't want to invest time to
it right now, so we're skipping those tests.

Signed-off-by: Joe Walker <jwalker@mozilla.com>