runat-1128988: Remote fixes for note and delegate types
Firstly the nodelist type tried to instansiate itself across the remote
interface, which was never going to work. Now it uses the 'remote' type
to proxy the type functions.

There is a special 'blank' conversion for types. parse() is async but
lots of the time we're effectively doing parse('') and making that case
async spreads the async virus to lots of places that it doesn't need to
be. So there's a getBlank() shortcut that is sync.

In almost all cases blank means 'INCOMPLETE' with an undefined value
(see [1] for the default implementation). But there are exceptions like
nodelist, where blank means an empty NodeList (which could be VALID)

So we add a blankIsValid property to the remote type which defaults to
false, so you get the default getBlank() behavior. But for nodelists
we set this to true, so we get the correct VALID / empty NodeList
behavior.

[1]: https://github.com/joewalker/gcli/blob/585f7b83953de5c0faeb9fb187b55ab6171c3d03/lib/gcli/types/types.js#L1007

Signed-off-by: Joe Walker <jwalker@mozilla.com>