runat-1128988: Replace windowHolder with environment
windowHolder was flawed in that there was only one per JS loader, but
we needed one for each command line.

The theory of 'environment' was that it wasn't owned by GCLI. It was
provided by the implementor of GCLI for the commands. This breaks that
by requiring that 'window' have a special value (if we're going to use
the node/nodelist/resource/javascript types) Given the complexities of
alternative solutions, this seems like an ok trade-off.

This commit is implementation only. Next we're going to fix tests.

Signed-off-by: Joe Walker <jwalker@mozilla.com>