runat-1128988: Simplify header injected into tests
We're (broadly) doing 2 things here. Partly we're causing all mock
commands to be loaded into the server process. We're still running the
test commands against the client process. This exercises the more of
the infrastructure with the test suite.

The changes to do that double the size of the test runner that we
inject into all the unit test files. Cut and paste is bad, so we
abstract this into a new function in helpers.js

This change fixes how we alter the test files as we copy them from the
GCLI tree to mozilla-central.

It's probably best to understand this looking at what happens to
central when we run this command.

Signed-off-by: Joe Walker <jwalker@mozilla.com>