runat-1128988: Merge getSelectionData into getSeletionLookup
Selection types have 2 ways to set the available options. 'data' allows
you to provide an array of strings for types when the underlying type
is a string. 'lookup' allows you to provide an array of objects with
name and value properties for times when the underlying type isn't a
string.

Previously we had 2 remote functions one for data and one for lookup.

The selection type has a function (getLookup) which does the work of
resolving promises, resolving cases where data or lookup is a function
and converting from data to lookup. There wasn't much point in having 2
remote functions when we could just use getLookup to resolve both.

The changes to remoted.js will need to be mirrored in gcli.js in
Firefox.

Signed-off-by: Joe Walker <jwalker@mozilla.com>