runat-1128988: Update test skips for remote testing
Lots of JS tests affected by how it's remoted, and the skips in testJs
were on crack, but that's unused in Firefox so we're not too worried
about that.

Resource tests are now run remotely too, and we can't run tests locally
on firefox. The new command test helps with that.

Signed-off-by: Joe Walker <jwalker@mozilla.com>